apps/renderer/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-pptx==0.6.23
Pillow==10.2.0
